+++ yangvoodoo/basedal.py
@@ -28,7 +28,7 @@ class BaseDataAbstractionLayer:
             log.setLevel(log_level)
         self.log = log
         self.session = None
-        self.conn = None
+        self.connected = False
         self.dirty = None
         self.datastore_changed = None
         self.module = None
@@ -204,7 +204,7 @@ class BaseDataAbstractionLayer:
     def has_datastore_changed(self):
         raise NotImplementedError("has_datastore_changed not implemented")
 
-    def dump_xpaths(self):
+    def dump_xpaths(self, start_xpath=None):
         raise NotImplementedError("dump_xpaths not implemented")
 
     def empty(self):