+++ yangvoodoo/stublydal.py
@@ -117,7 +117,19 @@ class StubLyDataAbstractionLayer(BaseDataAbstractionLayer):
         if not self.connected:
             raise NotConnect()
         self.log.trace("LIBYANG-GET: %s", xpath)
-        return next(self.libyang_data.get_xpath(xpath))
+        try:
+            return next(self.libyang_data.get_xpath(xpath))
+        except libyang.util.LibyangError:
+            return None
+
+    def libyang_gets_xpath(self, xpath):
+        """
+        Return a libyang-cffi DataNode directly, returning a generator of all entries that match
+        the XPATH.
+        """
+        if not self.connected:
+            raise Errors.NotConnect()
+        return self.libyang_data.get_xpath(xpath)
 
     def gets(self, xpath):
         if not self.connected:
@@ -231,17 +243,20 @@ class StubLyDataAbstractionLayer(BaseDataAbstractionLayer):
         self.log.trace("DELETE: %s", xpath)
         self.libyang_data.set_xpath(xpath, None)
 
-    def dump_xpaths(self):
-        """
-        This is not supported because libyang does not give us a convenient data
-        path. We can keep iterating around the children - but intermediate
-        nodes (like, bronze, silver, gold would by default get yielded as their
-        paths). We could add logic to find if they infact have values and skip
-        them- but that is more work and we have dump() available now.
-        """
+    def dump_xpaths(self, start_xpath=None):
         if not self.connected:
             raise NotConnect()
-        return {node.xpath: node.value for node in self.libyang_data.dump_datanodes()}
+        data_node = None
+        if start_xpath:
+            try:
+                data_node = next(self.libyang_data.get_xpath(start_xpath))
+            except StopIteration:
+                pass
+
+        return {
+            node.xpath: node.value
+            for node in self.libyang_data.dump_datanodes(start_node=data_node)
+        }
 
     def empty(self):
         raise NotImplementedError("empty not implemented")
@@ -270,7 +285,7 @@ class StubLyDataAbstractionLayer(BaseDataAbstractionLayer):
         self.log.trace("LOADS: (format: %s)", format)
         self.libyang_data.loads(payload, format, trusted)
 
-    def merges(self, filename, format=1, trusted=True):
+    def merge(self, filename, format=1, trusted=True):
         if not self.connected:
             raise NotConnect()
         self.log.trace("MERGE: (format: %s)", format)