+++ CHANGELOG.md
@@ -142,3 +142,20 @@
 
    - Rebuild docker images with new libyang __init__
    - Implement `libyang_get_xpath`
+
+# 0.0.9.2
+
+   - Allow dump_xpaths() to take in a starting xpath
+     (requires upgraded libyang-cfii)
+
+
+# 0.0.10
+
+   - Upgrade libyang to v1.0.240
+   - Utility function to `show_example_xpaths()`
+ 
+# 0.0.11
+  
+   - remove `data_abstraction_layer`- stublydal is the only option.
+   - add `libyang_gets_xpath` to return generator of DataNode's
+   - add `libyang_get_xpath` to return DataNode (first match or None)